Extract tool surface mode types into tool_surface/modes.rs
